Fixed all clippy warnings
a bunch of small things to make everything just that bit prettier

also replaced some old direction arrays I forgot...

+++ src/solutions/d9.rs
@@ -124,8 +124,8 @@ pub fn solve(input: Input) -> Output {
         let checksum = shrink_1(&mut fragmented_blocks);
         let unfragmented_checksum = shrink_2(&mut blocks);
 
-        total += checksum as usize;
-        unfragmented_total += unfragmented_checksum as usize;
+        total += checksum;
+        unfragmented_total += unfragmented_checksum;
     }
 
     output!(total, unfragmented_total)
